Move some code to odata library
