+++ routes/handlers/event.js
@@ -1,7 +1,7 @@
 const Event = require('../../models/osdi/event');
 const Joi = require('joi');
 const lodash = require('lodash');
-const createFilter = require('odata-v4-mongodb').createFilter;
+const ODATA = require('../../lib/odata');
 
 const OPTS_SCHEMA = Joi.object().keys({
   per_page: Joi.number().integer().min(1).default(25),
@@ -15,8 +15,8 @@ const OPTS_SCHEMA = Joi.object().keys({
 exports.get = function (opts, next) {
   Joi.validate(opts.query, OPTS_SCHEMA, function (err, query) {
     if (err) handleError(next, 'validating', err);
-    const filter = createFilter(query.$filter);
-    const orderBy = createOrderBy(query.$orderby);
+    const filter = ODATA.createFilter(query.$filter);
+    const orderBy = ODATA.createOrderBy(query.$orderby);
     console.log(`mongo db find = ${JSON.stringify(filter)} sort = ${JSON.stringify(orderBy)}`);
     Event.count(filter)
       .exec(function (err, count) {
@@ -108,36 +108,6 @@ exports.create = {
       });
   }
 };
-
-const createOrderBy = function (query) {
-  try {
-    if (query) {
-      var fields, direction;
-      const queryFields = query.trim().split(/[\s,]+/);
-      const lastField = queryFields[queryFields.length - 1];
-      if (lastField.toLowerCase() === 'asc') {
-        fields = queryFields.slice(0, -1);
-        direction = -1;
-      } else if (lastField.toLowerCase() === 'desc') {
-        fields = queryFields.slice(0, -1);
-        direction = 1;
-      } else {
-        fields = queryFields;
-        direction = -1;
-      }
-      const orderBy = [];
-      fields.forEach(function (field) {
-        orderBy.push([field, direction]);
-      });
-      return orderBy;
-    } else {
-      return [];
-    }
-  } catch (err) {
-    if (err) console.log(err);
-  }
-};
-exports.createOrderBy = createOrderBy;
 
 const handleError = function (next, str, err) {
   console.log(str, err);
